Revise attempt at using `/dev/gpiomem` by default.
I thought that wiringPi would only attempt to use `/dev/gpiomem` if
`WIRINGPI_GPIOMEM` == `1`, and therefore `setenv(x, 1, 0)` -- when set
to *not* overwrite an existing value -- would only == `1` if the user
had not explicitly set to some other value. While this is true, it ends
up that wiringPi only checks for the envvar to *exist* -- no matter what
the value. See:
https://git.drogon.net/?p=wiringPi;a=blob;f=wiringPi/wiringPi.c;hb=78b5c323b74de782df58ee558c249e4e4fadd25f#l1980

Therefore, I've changed it to specifically check for `0`, and try
`/dev/gpiomem`. If `0` or `/dev/gpiomem` fails, fall back to
`wiringPiSetupSys`, and if this fails, `return 1`.